Add firefly gan
